requirements.txt
Flask==1.1.2
tensorflow==2.11.0
numpy==1.21.5
Pillow==9.2.0
